+++ src/cmd/inventory.rs
@@ -118,9 +118,8 @@ async fn build(opts: &ConfigOpts) -> Result<()> {
             }
             // Get zone choice
             if let Some(idx) = scanner
-                .prompt("(Step 1 of 2) Choose a zone")
+                .prompt_t::<usize>("(Step 1 of 2) Choose a zone", "number")
                 .await?
-                .and_then(|s| s.parse::<usize>().ok())
             {
                 if idx > 0 && idx <= zones.len() {
                     break idx;
@@ -141,9 +140,11 @@ async fn build(opts: &ConfigOpts) -> Result<()> {
                     println!("[{}] {}", i + 1, record);
                 }
                 if let Some(idx) = scanner
-                    .prompt("(Step 2 of 2) Choose a record")
+                    .prompt_t::<usize>(
+                        "(Step 2 of 2) Choose a record",
+                        "number",
+                    )
                     .await?
-                    .and_then(|s| s.parse::<usize>().ok())
                 {
                     if idx > 0 && idx <= zone_records.len() {
                         break idx;
@@ -157,13 +158,13 @@ async fn build(opts: &ConfigOpts) -> Result<()> {
             let zone_id = selected_zone.id.clone();
             let record_id = selected_record.id.clone();
             inventory.insert(zone_id, record_id);
-            println!("\n✅ Added '{}'.", selected_record.name);
+            println!("✅ Added '{}'.", selected_record.name);
         } else {
-            println!("\n❌ No records for this zone.")
+            println!("❌ No records for this zone.")
         }
 
         let finished = 'finished: loop {
-            match scanner.prompt("Add another record? [Y/n]").await? {
+            match scanner.prompt("Add another record?", "Y/n").await? {
                 Some(input) => match input.to_lowercase().as_str() {
                     "y" | "yes" => break false,
                     "n" | "no" => break true,
@@ -179,10 +180,13 @@ async fn build(opts: &ConfigOpts) -> Result<()> {
 
     // Save
     let path = scanner
-        .prompt_path(format!(
-            "Save location [default: {}]",
-            default_inventory_path().display()
-        ))
+        .prompt_t::<PathBuf>(
+            format!(
+                "Save location, default: `{}`",
+                default_inventory_path().display()
+            ),
+            "path",
+        )
         .await?
         .map(|p| match p.extension() {
             Some(_) => p,
@@ -318,19 +322,13 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
         }
         // Ask to fix records
         let fix = force
-            || 'fix: loop {
-                match scanner
-                    .prompt(format!("Fix {} bad records? [Y/n]", bad.len()))
-                    .await?
-                {
-                    Some(input) => match input.to_lowercase().as_str() {
-                        "y" | "yes" => break true,
-                        "n" | "no" => break false,
-                        _ => continue 'fix,
-                    },
-                    None => break true,
-                }
-            };
+            || scanner
+                .prompt_yes_or_no(
+                    format!("Fix {} bad records?", bad.len()),
+                    "Y/n",
+                )
+                .await?
+                .unwrap_or(true);
         // Fix records
         let mut fixed = HashSet::new();
         if fix {
@@ -377,22 +375,13 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
         }
         // Ask to prune records
         let prune = force
-            || 'prune: loop {
-                match scanner
-                    .prompt(format!(
-                        "Prune {} invalid records? [Y/n]",
-                        invalid.len()
-                    ))
-                    .await?
-                {
-                    Some(input) => match input.to_lowercase().as_str() {
-                        "n" | "no" => break false,
-                        "y" | "yes" => break true,
-                        _ => continue 'prune,
-                    },
-                    None => break true,
-                }
-            };
+            || scanner
+                .prompt_yes_or_no(
+                    format!("Prune {} invalid records?", invalid.len()),
+                    "Y/n",
+                )
+                .await?
+                .unwrap_or(true);
         // Prune
         let mut pruned = HashSet::new();
         if prune {
@@ -444,11 +433,7 @@ pub async fn watch(opts: &ConfigOpts) -> Result<()> {
         opts.watch
             .as_ref()
             .and_then(|opts| opts.interval)
-            .unwrap_or_else(|| {
-                ConfigOptsWatch::default()
-                    .interval
-                    .expect("no default interval")
-            }),
+            .unwrap_or_else(|| ConfigOptsWatch::default().interval.unwrap()),
     );
 
     if interval.is_zero() {
@@ -469,6 +454,7 @@ pub async fn watch(opts: &ConfigOpts) -> Result<()> {
             {
                 println!("Error: {:?}", e);
             }
+            println!("Sleeping...");
         }
     }
 }