+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Subheadings to categorize changes are `added, changed, deprecated, removed, fixe
 
 ## Unreleased
 ### added
-- Placeholder
+- All configuration options are now built with `config build`
 ### changed
 - Improved readability of errors with `inventory watch`
 ### deprecated